firmware: fixed macro TO_SEC(). improved sunset effect

+++ firmware/common.h
@@ -54,6 +54,6 @@ typedef struct {
 #define ALARM_STOP_REQUEST	0x0400
 
 #define WDAY_TO_ID(x)		(x - 1)
-#define TO_SEC(x)		(60 * x)
+#define TO_SEC(x)		(60 * (uint16_t)x)
 
 #endif
